add: actions authorization

+++ app/Http/Controllers/TaskStatusController.php
@@ -8,12 +8,15 @@ use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 
 class TaskStatusController extends Controller
 {
-    public function show(): RedirectResponse
+    private const UNTOUCHABLE_STATUSES_ID = [1, 2, 3, 4];
+
+    public function show()
     {
-        return redirect()->route('task_statuses.index');
+        abort(404);
     }
 
     public function index(): Application|Factory|View
@@ -25,6 +28,7 @@ class TaskStatusController extends Controller
 
     public function create(): Factory|View|Application
     {
+        Gate::authorize('create', TaskStatus::class);
         $status = new TaskStatus();
 
         return view('task_statuses.create', compact('status'));
@@ -32,9 +36,12 @@ class TaskStatusController extends Controller
 
     public function store(TaskStatusRequest $request): RedirectResponse
     {
+        Gate::authorize('create', TaskStatus::class);
+
         $status = new TaskStatus();
         $status->fill($request->all());
         $status->save();
+
         flash(__('messages.flash.status.success.create'))->success();
 
         return redirect()->route('task_statuses.index');
@@ -42,11 +49,15 @@ class TaskStatusController extends Controller
 
     public function edit(TaskStatus $taskStatus): Factory|View|Application
     {
+        Gate::authorize('update', $taskStatus);
+
         return view('task_statuses.edit', compact('taskStatus'));
     }
 
     public function update(TaskStatusRequest $request, TaskStatus $taskStatus): RedirectResponse
     {
+        Gate::authorize('update', $taskStatus);
+
         $taskStatus->fill($request->all());
         $taskStatus->save();
         flash(__('messages.flash.status.success.update'))->success();
@@ -56,8 +67,17 @@ class TaskStatusController extends Controller
 
     public function destroy(TaskStatus $taskStatus): RedirectResponse
     {
-        $taskStatus->delete();
-        flash(__('messages.flash.status.success.delete'))->success();
+        Gate::authorize('delete', $taskStatus);
+
+        $exceptions = collect(self::UNTOUCHABLE_STATUSES_ID);
+
+        if (!$exceptions->contains($taskStatus->id)) {
+            $taskStatus->delete();
+
+            flash(__('messages.flash.status.success.delete'))->success();
+        } else {
+            flash(__('messages.flash.status.fail.delete'))->error();
+        }
 
         return redirect()->route('task_statuses.index');
     }
